Refactor the 'missing' command to pull CVE details from NIST data feeds.
Using published JSON data feeds from NIST (instead of cvedetails.com) allows for more reliable parsing of CVE details and removes the dependency on 'kub-at/php-simple-html-dom-parser'. Having a more reliable data source for CVE details will allow for further automation in the near future.

Since we now have a reliable way to parse CVE details, I've included a couple new attributes to the CVE check: lastModifiedDate and publishedDate. The values are not being used anywhere at the moment, but it might aid in pull-request review where CVE details have been modified. Also, the old CVE source commonly included those values in the summary.

Finally, I've changed the check.json 'threat' datatype from a string to float. I believe the float datatype is more appropriate, and the change was able to be made without having any compatibility issues with the scan logic. I believe there is a need to allow checks to be released prior to a threat value being assigned - in which case threat would be set to null.